refactor: changed int to BigDecimal for money, Lot service accepted only dto in parameters

+++ src/main/java/io/github/etorg/lot/internal/service/dto/MakeBidDto.java
@@ -1,7 +1,8 @@
 package io.github.etorg.lot.internal.service.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record MakeBidDto(UUID lotId, String currency, int value) {
+public record MakeBidDto(UUID lotId, String currency, BigDecimal value) {
 	
 }
